Rename tyre_models to tyres to prevent misconception
