+++ src/brim/bicycle/__init__.py
@@ -21,6 +21,6 @@ from brim.bicycle.front_frames import (
 from brim.bicycle.grounds import FlatGround, GroundBase
 from brim.bicycle.rear_frames import RearFrameBase, RigidRearFrame, RigidRearFrameMoore
 from brim.bicycle.stationary_bicycle import StationaryBicycle
-from brim.bicycle.tyre_models import NonHolonomicTyre, TyreBase
+from brim.bicycle.tyres import NonHolonomicTyre, TyreBase
 from brim.bicycle.wheels import KnifeEdgeWheel, ToroidalWheel, WheelBase
 from brim.bicycle.whipple_bicycle import WhippleBicycle, WhippleBicycleMoore
